# src/app/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DealVault API",
    "version": "1.0.0",
    "description": "Commodity deal room management platform API. All endpoints require authentication via NextAuth session unless noted otherwise."
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Development" }
  ],
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token"
      }
    },
...
  },
  "security": [{ "session": [] }],
  "paths": {
...
    "/api/profile": {
      "get": {
        "tags": ["Profile"],
        "summary": "Get current user profile with stats",
        "responses": { "200": { "description": "User profile with deal counts" } }
      },
      "patch": {
        "tags": ["Profile"],
        "summary": "Update profile (name, phone, password)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "phone": { "type": "string" },
                  "currentPassword": { "type": "string", "description": "Required when changing password" },
                  "newPassword": { "type": "string", "minLength": 8 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Profile updated" },
          "400": { "description": "Current password incorrect" }
        }
      }
    },
...
  }
}